linalg/wasm: start a wasm/ module tree with the activations, int8 kernel and madd macros
wasm.rs was a single 3159-line file, while every other linalg backend is a
module root plus a subdirectory. Move the three parts that carry no f32 MMM
kernel code into wasm/madd.rs, wasm/mmm_i32.rs and wasm/act.rs, and re-export
them from the root so every existing path still resolves. Pure relocation: the
emitted kernel bodies are unchanged under both +simd128 and +simd128,+relaxed-simd.
