+++ linalg/src/wasm.rs
@@ -14,50 +14,16 @@ use crate::{Ops, Scaler};
 #[cfg(target_feature = "relaxed-simd")]
 use crate::frame::element_wise::ElementWiseKer;
 
-// f32x4 mul+add → relaxed FMA when the build has +relaxed-simd, else explicit
-// mul+add. Lets the MMM kernels emit f32x4.relaxed_madd without duplicating
-// kernel source. Per PR #2199: LLVM does not auto-emit relaxed_madd from
-// f32x4_add(f32x4_mul(...)) even with +relaxed-simd — hand emission is needed.
-//
-// Caller must have `use std::arch::wasm32::*;` in scope (every kernel does).
-// Args are passed (acc, a, b); evaluation order differs between the two arms
-// (acc-first in baseline, acc-last in FMA), so callers must pass simple
-// variable names rather than expressions with side effects.
+#[macro_use]
+mod madd;
+
 #[cfg(target_feature = "relaxed-simd")]
-macro_rules! madd_f32x4 {
-    ($acc:expr, $a:expr, $b:expr) => {
-        f32x4_relaxed_madd($a, $b, $acc)
-    };
-}
+mod act;
+mod mmm_i32;
 
-#[cfg(not(target_feature = "relaxed-simd"))]
-macro_rules! madd_f32x4 {
-    ($acc:expr, $a:expr, $b:expr) => {
-        f32x4_add($acc, f32x4_mul($a, $b))
-    };
-}
-
-// Always-non-fused madd. Used by kernels with ≤4 SIMD accumulators per K-step
-// (wasm_f32_4x1, _8x1, _16x1, _4x4), where the destructive `fmla.4s`
-// emitted by +relaxed-simd creates a 4-cycle accumulator RAW recurrence
-// that throttles throughput to 1 FMA/cycle even though Apple-class ARM64
-// pipes can do 4. The separate `fmul.4s; fadd.4s` form gives each multiply
-// a fresh destination register, letting the OoO renamer overlap the next
-// iteration's multiply with the in-flight add. Measured: under
-// +simd128,+relaxed-simd these kernels are 19-28% slower than under
-// +simd128 when using the fused form on Apple M1 — both wasmtime
-// (Cranelift) and Node 20 (V8) reproduce identically. Wider kernels
-// (wasm_f32_32x1 with 8 accs, wasm_f32_8x8 with 16) keep the fused form
-// because their pipe is saturated and FMA's 1-instruction-per-madd wins.
-//
-// Cross-check: XNNPACK only ships wasmrelaxedsimd-fma GEMM kernels at
-// NR=8 (i.e. ≥8 accumulator-equivalents), independently arriving at the
-// same threshold without writing it down.
-macro_rules! madd_f32x4_nofma {
-    ($acc:expr, $a:expr, $b:expr) => {
-        f32x4_add($acc, f32x4_mul($a, $b))
-    };
-}
+#[cfg(target_feature = "relaxed-simd")]
+pub use act::*;
+pub use mmm_i32::*;
 
 pub fn plug(ops: &mut Ops) {
     ops.mmm_impls.push(wasm_f32_4x4.mmm());
@@ -2130,341 +2096,6 @@ unsafe fn kernel_f32_8x8(mut pnl: *const FusedKerSpec<f32>) -> isize {
 // strategize drops it and routes every GEMM onto the 32x1 GEMV kernel.
 MMMRustKernel!(kernel_f32_8x8 => wasm_f32_8x8<f32>(8,8)@(8,8) quality(ImplementationQuality::ManuallyOptimized));
 
-// Wasm SIMD int8 -> i32 matmul kernel (4x4). WASM's only integer dot
-// (i32x4.relaxed_dot_i8x16_i7x16) is non-deterministic for full i8 (its 2nd
-// operand is i7), so for a bit-exact kernel the AddMatMul K-loop uses widening
-// i8->i32 + i32x4 mul/add (an extmul/SMLAL-style outer product). The quant
-// epilogue + fuse ops reuse the bit-exact scalar path (q_scale/q_shr/q_shl),
-// which is O(MR*NR) and negligible vs the O(MR*NR*K) inner loop. Bit-identical
-// to generic_i32_4x4; selected for i8 matmul via its ManuallyOptimized quality
-// (WASM had no int8 matmul kernel — int8 fell back to the generic scalar one).
-#[inline(never)]
-unsafe fn kernel_i32_4x4(mut pnl: *const FusedKerSpec<i32>) -> isize {
-    use crate::ScaleShiftAndRound;
-    use std::arch::wasm32::*;
-    unsafe {
-        let mut ab = [[0i32; 4]; 4];
-        loop {
-            if pnl.is_null() {
-                break;
-            }
-            match *pnl {
-                FusedKerSpec::Done => break,
-                FusedKerSpec::Clear => ab = [[0i32; 4]; 4],
-                FusedKerSpec::LoadTile(col_major, _row_major) => {
-                    for row in 0..4 {
-                        for col in 0..4 {
-                            ab[row][col] = *col_major.add(col * 4 + row);
-                        }
-                    }
-                }
-                FusedKerSpec::ScalarAdd(a) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] += a;
-                        }
-                    }
-                }
-                FusedKerSpec::ScalarMul(a) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] *= a;
-                        }
-                    }
-                }
-                FusedKerSpec::ScalarMin(m) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].min(m);
-                        }
-                    }
-                }
-                FusedKerSpec::ScalarMax(m) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].max(m);
-                        }
-                    }
-                }
-                FusedKerSpec::ScalarSub(m) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = m - ab[i][j];
-                        }
-                    }
-                }
-                FusedKerSpec::ScalarSubF(m) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] -= m;
-                        }
-                    }
-                }
-                FusedKerSpec::LeakyRelu(a) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = if ab[i][j] > 0 { ab[i][j] } else { a * ab[i][j] };
-                        }
-                    }
-                }
-                FusedKerSpec::PerRowMin(m) => {
-                    for i in 0..4 {
-                        let v = *m.add(i);
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].min(v);
-                        }
-                    }
-                }
-                FusedKerSpec::PerRowMax(m) => {
-                    for i in 0..4 {
-                        let v = *m.add(i);
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].max(v);
-                        }
-                    }
-                }
-                FusedKerSpec::PerRowAdd(m) => {
-                    for i in 0..4 {
-                        let v = *m.add(i);
-                        for j in 0..4 {
-                            ab[i][j] += v;
-                        }
-                    }
-                }
-                FusedKerSpec::PerRowMul(m) => {
-                    for i in 0..4 {
-                        let v = *m.add(i);
-                        for j in 0..4 {
-                            ab[i][j] *= v;
-                        }
-                    }
-                }
-                FusedKerSpec::PerRowSub(m) => {
-                    for i in 0..4 {
-                        let v = *m.add(i);
-                        for j in 0..4 {
-                            ab[i][j] = v - ab[i][j];
-                        }
-                    }
-                }
-                FusedKerSpec::PerRowSubF(m) => {
-                    for i in 0..4 {
-                        let v = *m.add(i);
-                        for j in 0..4 {
-                            ab[i][j] -= v;
-                        }
-                    }
-                }
-                FusedKerSpec::PerColMin(m) => {
-                    let c = std::slice::from_raw_parts(m, 4);
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].min(c[j]);
-                        }
-                    }
-                }
-                FusedKerSpec::PerColMax(m) => {
-                    let c = std::slice::from_raw_parts(m, 4);
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].max(c[j]);
-                        }
-                    }
-                }
-                FusedKerSpec::PerColAdd(m) => {
-                    let c = std::slice::from_raw_parts(m, 4);
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] += c[j];
-                        }
-                    }
-                }
-                FusedKerSpec::PerColMul(m) => {
-                    let c = std::slice::from_raw_parts(m, 4);
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] *= c[j];
-                        }
-                    }
-                }
-                FusedKerSpec::PerColSub(m) => {
-                    let c = std::slice::from_raw_parts(m, 4);
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = c[j] - ab[i][j];
-                        }
-                    }
-                }
-                FusedKerSpec::PerColSubF(m) => {
-                    let c = std::slice::from_raw_parts(m, 4);
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] -= c[j];
-                        }
-                    }
-                }
-                FusedKerSpec::AddRowColProducts(rows, cols) => {
-                    for i in 0..4 {
-                        let r = *rows.add(i);
-                        for j in 0..4 {
-                            ab[i][j] += r * *cols.add(j);
-                        }
-                    }
-                }
-                FusedKerSpec::AddUnicast(other) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            let p = other.ptr.offset(
-                                other.row_byte_stride * i as isize
-                                    + other.col_byte_stride * j as isize,
-                            );
-                            let v = match other.item_size {
-                                1 => *(p as *const i8) as i32,
-                                4 => *(p as *const i32),
-                                _ => return 1,
-                            };
-                            ab[i][j] += v;
-                        }
-                    }
-                }
-                FusedKerSpec::ShiftLeft(shift) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].q_shl(shift);
-                        }
-                    }
-                }
-                FusedKerSpec::RoundingShiftRight(shift, rp) => {
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].q_shr(shift, rp);
-                        }
-                    }
-                }
-                FusedKerSpec::QScale(shift, rp, mult) => {
-                    let s = Scaler::from_fuse_params(shift, rp, mult);
-                    for i in 0..4 {
-                        for j in 0..4 {
-                            ab[i][j] = ab[i][j].q_scale(s);
-                        }
-                    }
-                }
-                FusedKerSpec::AddMatMul { k, pa, pb, packing } => {
-                    if packing == 1 {
-                        let a = pa as *const i8;
-                        let b = pb as *const i8;
-                        let mut acc = [
-                            v128_load(ab[0].as_ptr() as *const v128),
-                            v128_load(ab[1].as_ptr() as *const v128),
-                            v128_load(ab[2].as_ptr() as *const v128),
-                            v128_load(ab[3].as_ptr() as *const v128),
-                        ];
-                        // PackedI8K4 (K=4-inner): per 4-K block, one B v128 load is
-                        // shared across the 4 rows; each row broadcasts its 4 K bytes
-                        // (a[kb*16 + m*4 ..]) and issues one relaxed_dot (16 MACs).
-                        // b[kb*16 + n*4 + kr]. Tail K (k%4) is zero-padded by the packer.
-                        #[cfg(target_feature = "relaxed-simd")]
-                        for kb in 0..k.div_ceil(4) {
-                            let b_all = v128_load(b.add(kb * 16) as *const v128);
-                            for (m, acc_m) in acc.iter_mut().enumerate() {
-                                let a4 = (a.add(kb * 16 + m * 4) as *const i32).read_unaligned();
-                                *acc_m = i32x4_relaxed_dot_i8x16_i7x16_add(
-                                    i32x4_splat(a4),
-                                    b_all,
-                                    *acc_m,
-                                );
-                            }
-                        }
-                        // Deterministic fallback (no relaxed-simd): standard PackedFormat
-                        // K-major (4 i8 per k), widening outer-product accumulate.
-                        #[cfg(not(target_feature = "relaxed-simd"))]
-                        for ik in 0..k {
-                            let bw = v128_load32_zero(b.add(4 * ik) as *const u32);
-                            let bw = i16x8_extend_low_i8x16(bw);
-                            let bw = i32x4_extend_low_i16x8(bw);
-                            let ar = a.add(4 * ik);
-                            acc[0] =
-                                i32x4_add(acc[0], i32x4_mul(i32x4_splat(*ar.add(0) as i32), bw));
-                            acc[1] =
-                                i32x4_add(acc[1], i32x4_mul(i32x4_splat(*ar.add(1) as i32), bw));
-                            acc[2] =
-                                i32x4_add(acc[2], i32x4_mul(i32x4_splat(*ar.add(2) as i32), bw));
-                            acc[3] =
-                                i32x4_add(acc[3], i32x4_mul(i32x4_splat(*ar.add(3) as i32), bw));
-                        }
-                        v128_store(ab[0].as_mut_ptr() as *mut v128, acc[0]);
-                        v128_store(ab[1].as_mut_ptr() as *mut v128, acc[1]);
-                        v128_store(ab[2].as_mut_ptr() as *mut v128, acc[2]);
-                        v128_store(ab[3].as_mut_ptr() as *mut v128, acc[3]);
-                    } else if packing == 0 {
-                        // i32 x i32, K-major (scalar; rare path).
-                        let a = pa as *const i32;
-                        let b = pb as *const i32;
-                        for ik in 0..k {
-                            for i in 0..4 {
-                                let av = *a.add(4 * ik + i);
-                                for j in 0..4 {
-                                    ab[i][j] += av * *b.add(4 * ik + j);
-                                }
-                            }
-                        }
-                    } else {
-                        return 1;
-                    }
-                }
-                FusedKerSpec::Store(tile) => match tile.item_size {
-                    1 => {
-                        for i in 0..4 {
-                            for j in 0..4 {
-                                let loc = tile.ptr.offset(
-                                    tile.row_byte_stride * i as isize
-                                        + tile.col_byte_stride * j as isize,
-                                ) as *mut u8;
-                                *loc = ab[i][j] as u8;
-                            }
-                        }
-                    }
-                    4 => {
-                        for i in 0..4 {
-                            for j in 0..4 {
-                                let loc = tile.ptr.offset(
-                                    tile.row_byte_stride * i as isize
-                                        + tile.col_byte_stride * j as isize,
-                                ) as *mut i32;
-                                *loc = ab[i][j];
-                            }
-                        }
-                    }
-                    _ => return 1,
-                },
-            };
-            pnl = pnl.add(1);
-        }
-    }
-    0
-}
-
-// i8i8 packing for wasm_i32_4x4. Under +relaxed-simd the kernel uses the
-// `i32x4_relaxed_dot_i8x16_i7x16_add` SDOT-analog, which wants 4 contiguous K per
-// mn-lane → PackedI8K4 (K=4-inner). Without relaxed-simd it uses the widening
-// outer-product, which wants K-major → standard PackedFormat. Both are picked at
-// compile time so the kernel's AddMatMul and the packer always agree.
-#[cfg(target_feature = "relaxed-simd")]
-fn wasm_i8_packing() -> impl crate::mmm::MMMInputFormat {
-    crate::pack::PackedI8K4::new(4)
-}
-#[cfg(not(target_feature = "relaxed-simd"))]
-fn wasm_i8_packing() -> impl crate::mmm::MMMInputFormat {
-    use crate::pack::Packing;
-    i8::packing(4)
-}
-
-MMMRustKernel!(kernel_i32_4x4 => wasm_i32_4x4<i32>(4,4)
-    packing[1] = i8i8 => |k| k.with_packing(wasm_i8_packing(), wasm_i8_packing());
-    quality(ImplementationQuality::ManuallyOptimized)
-    store(i8)
-);
-
 #[cfg(test)]
 mod dispatch_trace {
     fn trace_one(label: &str, m: Option<usize>, k: Option<usize>, n: Option<usize>) {
@@ -2881,279 +2512,5 @@ mod microbench_dispatch_gemv {
         bench_shape("M=64  k=96 ", 64, 96, 20_000);
         bench_shape("M=100 k=256", 100, 256, 10_000);
         bench_shape("M=256 k=256", 256, 256, 5_000);
-    }
-}
-
-// Relaxed-SIMD activation kernels (f32, FMA path).
-//
-// `f32x4_relaxed_madd(a, b, c)` computes `a * b + c`. On hosts with hardware
-// FMA (all ARM64, x86_64 with FMA3) it lowers to a single fused, single-
-// rounded instruction. On hosts without, it falls back to mul+add — hence
-// "relaxed". The result is therefore not bit-deterministic across all hosts,
-// but it is at least as accurate as the separate mul+add (FMA does fewer
-// roundings).
-//
-// For sigmoid/tanh polynomial evaluation, the 14 muladds in the Horner chain
-// fuse cleanly. Measured ~1.65x over the baseline-simd128 explicit kernel and
-// over LLVM auto-vec'd scalar on V8.
-//
-// Gated on `target_feature = "relaxed-simd"` because `f32x4_relaxed_madd`
-// requires the relaxed-simd proposal to be enabled at compile time.
-// ---------------------------------------------------------------------------
-
-#[cfg(target_feature = "relaxed-simd")]
-#[derive(Clone, Debug)]
-pub struct WasmSigmoid4Relaxed;
-
-#[cfg(target_feature = "relaxed-simd")]
-impl ElementWiseKer<f32> for WasmSigmoid4Relaxed {
-    fn name() -> &'static str {
-        "wasm_relaxed_simd"
-    }
-
-    fn alignment_bytes() -> usize {
-        16
-    }
-
-    fn alignment_items() -> usize {
-        4
-    }
-
-    fn nr() -> usize {
-        4
-    }
-
-    fn run(buf: &mut [f32], _: ()) {
-        use std::arch::wasm32::*;
-
-        debug_assert!(buf.len() % Self::nr() == 0);
-        debug_assert!(buf.as_ptr() as usize % Self::alignment_bytes() == 0);
-
-        // Coefficients match generic/sigmoid.rs::ssigmoid bit-for-bit.
-        // Output may differ by ≤1 ulp from scalar on FMA hosts (more accurate).
-        const LOW: f32 = -18.6;
-        const HIGH: f32 = -LOW;
-
-        const ALPHA_13: f32 = -4.433153405e-18;
-        const ALPHA_11: f32 = 1.169974371e-14;
-        const ALPHA_9: f32 = -1.875289645e-11;
-        const ALPHA_7: f32 = 4.257889523e-8;
-        const ALPHA_5: f32 = 0.00004811817576;
-        const ALPHA_3: f32 = 0.008163842030;
-        const ALPHA_1: f32 = 0.2499999971;
-
-        const BETA_6: f32 = 3.922935744e-6;
-        const BETA_4: f32 = 0.001524872358;
-        const BETA_2: f32 = 0.1159886749;
-        const BETA_0: f32 = 1.0;
-
-        unsafe {
-            let lo = f32x4_splat(LOW);
-            let hi = f32x4_splat(HIGH);
-
-            let a13 = f32x4_splat(ALPHA_13);
-            let a11 = f32x4_splat(ALPHA_11);
-            let a9 = f32x4_splat(ALPHA_9);
-            let a7 = f32x4_splat(ALPHA_7);
-            let a5 = f32x4_splat(ALPHA_5);
-            let a3 = f32x4_splat(ALPHA_3);
-            let a1 = f32x4_splat(ALPHA_1);
-
-            let b6 = f32x4_splat(BETA_6);
-            let b4 = f32x4_splat(BETA_4);
-            let b2 = f32x4_splat(BETA_2);
-            let b0 = f32x4_splat(BETA_0);
-
-            let half = f32x4_splat(0.5);
-
-            let mut p = buf.as_mut_ptr();
-            let end = p.add(buf.len());
-            while p < end {
-                let v = v128_load(p as *const v128);
-                let x = f32x4_min(hi, f32x4_max(lo, v));
-                let x2 = f32x4_mul(x, x);
-
-                // Horner numerator with FMA: pn = x2 * pn + a_n
-                let pn = a13;
-                let pn = f32x4_relaxed_madd(x2, pn, a11);
-                let pn = f32x4_relaxed_madd(x2, pn, a9);
-                let pn = f32x4_relaxed_madd(x2, pn, a7);
-                let pn = f32x4_relaxed_madd(x2, pn, a5);
-                let pn = f32x4_relaxed_madd(x2, pn, a3);
-                let pn = f32x4_relaxed_madd(x2, pn, a1);
-                let pn = f32x4_mul(pn, x);
-
-                // Horner denominator with FMA
-                let qn = b6;
-                let qn = f32x4_relaxed_madd(x2, qn, b4);
-                let qn = f32x4_relaxed_madd(x2, qn, b2);
-                let qn = f32x4_relaxed_madd(x2, qn, b0);
-
-                let r = f32x4_add(f32x4_div(pn, qn), half);
-                v128_store(p as *mut v128, r);
-                p = p.add(4);
-            }
-        }
-    }
-}
-
-#[cfg(target_feature = "relaxed-simd")]
-#[derive(Clone, Debug)]
-pub struct WasmTanh4Relaxed;
-
-#[cfg(target_feature = "relaxed-simd")]
-impl ElementWiseKer<f32> for WasmTanh4Relaxed {
-    fn name() -> &'static str {
-        "wasm_relaxed_simd"
-    }
-
-    fn alignment_bytes() -> usize {
-        16
-    }
-
-    fn alignment_items() -> usize {
-        4
-    }
-
-    fn nr() -> usize {
-        4
-    }
-
-    fn run(buf: &mut [f32], _: ()) {
-        use std::arch::wasm32::*;
-
-        debug_assert!(buf.len() % Self::nr() == 0);
-        debug_assert!(buf.as_ptr() as usize % Self::alignment_bytes() == 0);
-
-        const LOW: f32 = -8.9;
-        const HIGH: f32 = 8.9;
-
-        const ALPHA_13: f32 = -8.488492677e-14;
-        const ALPHA_11: f32 = 5.277853000e-11;
-        const ALPHA_9: f32 = -2.022500419e-8;
-        const ALPHA_7: f32 = 0.00001115424833;
-        const ALPHA_5: f32 = 0.003103950131;
-        const ALPHA_3: f32 = 0.1308400453;
-        const ALPHA_1: f32 = 0.9999999934;
-
-        const BETA_6: f32 = 0.0002546136580;
-        const BETA_4: f32 = 0.02449515379;
-        const BETA_2: f32 = 0.4641733162;
-        const BETA_0: f32 = 1.0;
-
-        unsafe {
-            let lo = f32x4_splat(LOW);
-            let hi = f32x4_splat(HIGH);
-
-            let a13 = f32x4_splat(ALPHA_13);
-            let a11 = f32x4_splat(ALPHA_11);
-            let a9 = f32x4_splat(ALPHA_9);
-            let a7 = f32x4_splat(ALPHA_7);
-            let a5 = f32x4_splat(ALPHA_5);
-            let a3 = f32x4_splat(ALPHA_3);
-            let a1 = f32x4_splat(ALPHA_1);
-
-            let b6 = f32x4_splat(BETA_6);
-            let b4 = f32x4_splat(BETA_4);
-            let b2 = f32x4_splat(BETA_2);
-            let b0 = f32x4_splat(BETA_0);
-
-            let mut p = buf.as_mut_ptr();
-            let end = p.add(buf.len());
-            while p < end {
-                let v = v128_load(p as *const v128);
-                let x = f32x4_min(hi, f32x4_max(lo, v));
-                let x2 = f32x4_mul(x, x);
-
-                let pn = a13;
-                let pn = f32x4_relaxed_madd(x2, pn, a11);
-                let pn = f32x4_relaxed_madd(x2, pn, a9);
-                let pn = f32x4_relaxed_madd(x2, pn, a7);
-                let pn = f32x4_relaxed_madd(x2, pn, a5);
-                let pn = f32x4_relaxed_madd(x2, pn, a3);
-                let pn = f32x4_relaxed_madd(x2, pn, a1);
-                let pn = f32x4_mul(pn, x);
-
-                let qn = b6;
-                let qn = f32x4_relaxed_madd(x2, qn, b4);
-                let qn = f32x4_relaxed_madd(x2, qn, b2);
-                let qn = f32x4_relaxed_madd(x2, qn, b0);
-
-                let r = f32x4_div(pn, qn);
-                v128_store(p as *mut v128, r);
-                p = p.add(4);
-            }
-        }
-    }
-}
-
-#[cfg(all(test, target_feature = "relaxed-simd"))]
-#[macro_use]
-mod test_wasm_sigmoid_relaxed {
-    sigmoid_frame_tests!(true, f32, crate::wasm::WasmSigmoid4Relaxed);
-}
-
-#[cfg(all(test, target_feature = "relaxed-simd"))]
-#[macro_use]
-mod test_wasm_tanh_relaxed {
-    tanh_frame_tests!(true, f32, crate::wasm::WasmTanh4Relaxed);
-}
-
-#[cfg(all(test, target_feature = "relaxed-simd"))]
-mod microbench_activations {
-    //! Microbench: WASM SIMD sigmoid/tanh vs the generic scalar fallback.
-    //! Sizes mirror typical RNN/transformer hidden dims (256, 512, 1024).
-    //!
-    //! Run with:
-    //!   RUSTFLAGS='-C target-feature=+simd128' \
-    //!     CARGO_TARGET_WASM32_WASIP1_RUNNER='wasmtime --env RUST_TEST_NOCAPTURE=1 --' \
-    //!     cargo test --release --target=wasm32-wasip1 -p tract-linalg \
-    //!     wasm::microbench_activations::microbench -- --nocapture --ignored
-    use crate::frame::element_wise::ElementWiseKer;
-    use std::time::Instant;
-
-    fn ns_per_call<K: ElementWiseKer<f32>>(buf: &mut [f32], iters: usize) -> f64 {
-        // Warmup
-        for _ in 0..50 {
-            K::run(buf, ());
-        }
-        let t0 = Instant::now();
-        for _ in 0..iters {
-            K::run(buf, ());
-        }
-        let elapsed = t0.elapsed();
-        elapsed.as_secs_f64() / iters as f64 * 1e9
-    }
-
-    fn bench(label: &str, n: usize, iters: usize) {
-        // Same input for both kernels — rebuild between to avoid post-clamp
-        // saturation mucking up the measurement.
-        let make = || (0..n).map(|i| ((i % 37) as f32 - 18.0) * 0.5).collect::<Vec<f32>>();
-
-        let mut buf = make();
-        let scalar_sig = ns_per_call::<crate::generic::sigmoid::SSigmoid4>(&mut buf, iters);
-        let mut buf = make();
-        let simd_sig = ns_per_call::<crate::wasm::WasmSigmoid4Relaxed>(&mut buf, iters);
-        let mut buf = make();
-        let scalar_tanh = ns_per_call::<crate::generic::tanh::STanh4>(&mut buf, iters);
-        let mut buf = make();
-        let simd_tanh = ns_per_call::<crate::wasm::WasmTanh4Relaxed>(&mut buf, iters);
-
-        eprintln!(
-            "{label} n={n} iters={iters}: \
-             sigmoid scalar={scalar_sig:.0} ns simd={simd_sig:.0} ns ({:.2}x); \
-             tanh scalar={scalar_tanh:.0} ns simd={simd_tanh:.0} ns ({:.2}x)",
-            scalar_sig / simd_sig,
-            scalar_tanh / simd_tanh,
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn microbench() {
-        eprintln!("=== WASM SIMD activations: scalar vs simd ===");
-        bench("hidden=256", 256, 5_000);
-        bench("hidden=512", 512, 3_000);
-        bench("hidden=1024", 1024, 2_000);
     }
 }
